# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/companies": {
      "get": {
        "operationId": "CompaniesController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Поиск по названию компании",
            "schema": {
              "example": "google",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCompaniesResponseDto"
                }
              }
            }
          }
        },
        "summary": "Список компаний",
        "tags": [
          "companies"
        ]
      },
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              }
            }
          }
        },
        "summary": "Создать компанию",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/companies/{id}": {
      "patch": {
        "operationId": "CompaniesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              }
            }
          }
        },
        "summary": "Переименовать компанию",
        "tags": [
          "companies"
        ]
      }
    },
...
    "/api/interviews/cycles": {
      "post": {
        "operationId": "InterviewsController_createCycle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewCycleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewCycleDetailResponseDto"
                }
              }
            }
          }
        },
        "summary": "Создать weekly cycle и автосгенерировать draft-пары",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/interviews/cycles/{id}": {
      "get": {
        "operationId": "InterviewsController_getCycleDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewCycleDetailResponseDto"
                }
              }
            }
          }
        },
        "summary": "Детали interview cycle",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/interviews/cycles/{id}/pairs": {
      "post": {
        "operationId": "InterviewsController_createPair",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewPairDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewItemDto"
                }
              }
            }
          }
        },
        "summary": "Ручное создание пары в cycle",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/interviews/{id}": {
      "patch": {
        "operationId": "InterviewsController_updateInterview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInterviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewItemDto"
                }
              }
            }
          }
        },
        "summary": "Обновить интервью",
        "tags": [
          "interviews"
        ]
      },
      "delete": {
        "operationId": "InterviewsController_removeInterview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Удалить интервью",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/interviews/admin-calendar": {
      "get": {
        "operationId": "InterviewsController_getAdminCalendar",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "description": "По умолчанию текущий UTC-месяц",
            "schema": {
              "example": "2026-03",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminInterviewCalendarResponseDto"
                }
              }
            }
          }
        },
        "summary": "Админский month calendar интервью",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/interviews/my-calendar": {
      "get": {
        "operationId": "InterviewsController_getMyCalendar",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "description": "По умолчанию текущий UTC-месяц",
            "schema": {
              "example": "2026-03",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyInterviewCalendarResponseDto"
                }
              }
            }
          }
        },
        "summary": "Личный month calendar интервью",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/interviews/{id}/runtime": {
      "get": {
        "operationId": "InterviewsController_getRuntime",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewRuntimeResponseDto"
                }
              }
            }
          }
        },
        "summary": "Runtime payload интервью для interviewer",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/interviews/{id}/complete": {
      "post": {
        "operationId": "InterviewsController_completeInterview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteInterviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewItemDto"
                }
              }
            }
          }
        },
        "summary": "Завершить интервью как interviewer",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/interviews/admin-dashboard": {
      "get": {
        "operationId": "InterviewsController_getAdminDashboard",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-02-16",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-04-12",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminInterviewDashboardResponseDto"
                }
              }
            }
          }
        },
        "summary": "Admin dashboard по интервью",
        "tags": [
          "interviews"
        ]
      }
    },
    "/api/interviews/my-dashboard": {
      "get": {
        "operationId": "InterviewsController_getMyDashboard",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-02-16",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "example": "2026-04-12",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyInterviewDashboardResponseDto"
                }
              }
            }
          }
        },
        "summary": "User dashboard по интервью",
        "tags": [
          "interviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nord API",
    "description": "Backend API for interview questions bank",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CompanyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm7y5z8qv0003x4f2w7sn1abc"
          },
          "name": {
            "type": "string",
            "example": "Google"
          },
          "questionsCount": {
            "type": "number",
            "example": 12
          }
        },
        "required": [
          "id",
          "name",
          "questionsCount"
        ]
      },
...
      "ListCompaniesMetaDto": {
        "type": "object",
        "properties": {
          "tookMs": {
            "type": "number",
            "example": 9
          },
          "appliedFilters": {
            "$ref": "#/components/schemas/ListCompaniesAppliedFiltersDto"
          }
        },
        "required": [
          "tookMs",
          "appliedFilters"
        ]
      },
      "ListCompaniesResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyDto"
            }
          },
          "total": {
            "type": "number",
            "example": 15
          },
          "meta": {
            "$ref": "#/components/schemas/ListCompaniesMetaDto"
          }
        },
        "required": [
          "items",
          "total",
          "meta"
        ]
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название компании",
            "maxLength": 100,
            "example": "Google"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Новое название компании",
            "maxLength": 100,
            "example": "Google Cloud"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateInterviewCycleDto": {
        "type": "object",
        "properties": {
          "periodStart": {
            "type": "string",
            "format": "date",
            "example": "2026-03-09"
          },
          "periodEnd": {
            "type": "string",
            "format": "date",
            "example": "2026-03-15"
          },
          "participantIds": {
            "example": [
              "cm8user1",
              "cm8user2",
              "cm8user3"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "periodStart",
          "periodEnd",
          "participantIds"
        ]
      },
      "InterviewUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm8q4x7r10001s2t3u4v5w6x7"
          },
          "login": {
            "type": "string",
            "example": "ivan.petrov"
          },
          "displayName": {
            "type": "string",
            "example": "Иван Петров"
          }
        },
        "required": [
          "id",
          "login",
          "displayName"
        ]
      },
      "InterviewPresetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm8q4x7r10001preset"
          },
          "name": {
            "type": "string",
            "example": "Angular Developer"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "InterviewItemDto": {
        "type": "object",
        "properties": {
          "resultsCount": {
            "type": "number",
            "example": 10
          },
          "correctCount": {
            "type": "number",
            "example": 4
          },
          "incorrectCount": {
            "type": "number",
            "example": 3
          },
          "partialCount": {
            "type": "number",
            "example": 3
          },
          "id": {
            "type": "string",
            "example": "cm8q4x7r10001interview"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PLANNED",
              "SCHEDULED",
              "COMPLETED"
            ],
            "example": "SCHEDULED"
          },
          "plannedDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-03-11"
          },
          "preset": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewPresetDto"
              }
            ]
          },
          "interviewer": {
            "$ref": "#/components/schemas/InterviewUserDto"
          },
          "interviewee": {
            "$ref": "#/components/schemas/InterviewUserDto"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "resultsCount",
          "correctCount",
          "incorrectCount",
          "partialCount",
          "id",
          "status",
          "plannedDate",
          "preset",
          "interviewer",
          "interviewee",
          "completedAt"
        ]
      },
      "InterviewCycleDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm8q4x7r10001cycle"
          },
          "mode": {
            "type": "string",
            "enum": [
              "AUTO",
              "MANUAL"
            ],
            "example": "AUTO"
          },
          "periodStart": {
            "type": "string",
            "format": "date",
            "example": "2026-03-09"
          },
          "periodEnd": {
            "type": "string",
            "format": "date",
            "example": "2026-03-15"
          },
          "createdByAdmin": {
            "$ref": "#/components/schemas/InterviewUserDto"
          },
          "interviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewItemDto"
            }
          },
          "draftCount": {
            "type": "number",
            "example": 2
          },
          "plannedCount": {
            "type": "number",
            "example": 1
          },
          "scheduledCount": {
            "type": "number",
            "example": 3
          },
          "completedCount": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "id",
          "mode",
          "periodStart",
          "periodEnd",
          "createdByAdmin",
          "interviews",
          "draftCount",
          "plannedCount",
          "scheduledCount",
          "completedCount"
        ]
      },
      "CreateInterviewPairDto": {
        "type": "object",
        "properties": {
          "interviewerId": {
            "type": "string",
            "example": "cm8user1"
          },
          "intervieweeId": {
            "type": "string",
            "example": "cm8user2"
          }
        },
        "required": [
          "interviewerId",
          "intervieweeId"
        ]
      },
      "UpdateInterviewDto": {
        "type": "object",
        "properties": {
          "interviewerId": {
            "type": "object",
            "example": "cm8user1",
            "nullable": true
          },
          "intervieweeId": {
            "type": "object",
            "example": "cm8user2",
            "nullable": true
          },
          "plannedDate": {
            "type": "string",
            "format": "date",
            "example": "2026-03-11",
            "nullable": true
          },
          "presetId": {
            "type": "object",
            "example": "cm8preset1",
            "nullable": true
          },
          "feedback": {
            "type": "object",
            "example": "Нужно точнее формулировать компромиссы.",
            "nullable": true
          }
        }
      },
      "InterviewCalendarDayDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "example": "2026-03-11"
          }
        },
        "required": [
          "date"
        ]
      },
      "InterviewCycleDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm8q4x7r10001cycle"
          },
          "mode": {
            "type": "string",
            "enum": [
              "AUTO",
              "MANUAL"
            ],
            "example": "AUTO"
          },
          "periodStart": {
            "type": "string",
            "format": "date",
            "example": "2026-03-09"
          },
          "periodEnd": {
            "type": "string",
            "format": "date",
            "example": "2026-03-15"
          },
          "createdByAdmin": {
            "$ref": "#/components/schemas/InterviewUserDto"
          },
          "interviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewItemDto"
            }
          }
        },
        "required": [
          "id",
          "mode",
          "periodStart",
          "periodEnd",
          "createdByAdmin",
          "interviews"
        ]
      },
      "AdminInterviewCalendarResponseDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "example": "2026-03"
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewCalendarDayDto"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewItemDto"
            }
          },
          "activeCycle": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewCycleDetailDto"
              }
            ]
          }
        },
        "required": [
          "month",
          "days",
          "items",
          "activeCycle"
        ]
      },
...
      "MyInterviewCalendarResponseDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "example": "2026-03"
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewCalendarDayDto"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyInterviewCalendarItemDto"
            }
          }
        },
        "required": [
          "month",
          "days",
          "items"
        ]
      },
...
      "InterviewRuntimeResponseDto": {
        "type": "object",
        "properties": {
          "interview": {
            "$ref": "#/components/schemas/InterviewItemDto"
          },
          "counterpart": {
            "$ref": "#/components/schemas/InterviewUserDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewRuntimeQuestionDto"
            }
          }
        },
        "required": [
          "interview",
          "counterpart",
          "items"
        ]
      },
      "CompleteInterviewItemDto": {
        "type": "object",
        "properties": {
          "interviewQuestionId": {
            "type": "string",
            "example": "cm8interviewquestion1"
          },
          "result": {
            "type": "string",
            "enum": [
              "correct",
              "incorrect",
              "partial"
            ],
            "example": "partial"
          }
        },
        "required": [
          "interviewQuestionId",
          "result"
        ]
      },
      "CompleteInterviewDto": {
        "type": "object",
        "properties": {
          "feedback": {
            "type": "string",
            "example": "Нужно точнее формулировать компромиссы и быстрее структурировать ответ."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompleteInterviewItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
      "InterviewDashboardOutcomeMixDto": {
        "type": "object",
        "properties": {
          "correctCount": {
            "type": "number",
            "example": 12
          },
          "partialCount": {
            "type": "number",
            "example": 8
          },
          "incorrectCount": {
            "type": "number",
            "example": 5
          }
        },
        "required": [
          "correctCount",
          "partialCount",
          "incorrectCount"
        ]
      },
...
      "AdminInterviewDashboardResponseDto": {
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/InterviewDashboardSummaryDto"
          },
          "scheduleSeries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewDashboardScheduleSeriesItemDto"
            }
          },
          "outcomeMix": {
            "$ref": "#/components/schemas/InterviewDashboardOutcomeMixDto"
          },
          "interviewerLoad": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewDashboardInterviewerLoadItemDto"
            }
          },
          "weakTopics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewWeakTopicDto"
            }
          },
          "upcoming": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewDashboardRecentItemDto"
            }
          },
          "recentCompleted": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewDashboardRecentItemDto"
            }
          }
        },
        "required": [
          "summary",
          "scheduleSeries",
          "outcomeMix",
          "interviewerLoad",
          "weakTopics",
          "upcoming",
          "recentCompleted"
        ]
      },
...
      "MyInterviewDashboardResponseDto": {
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/InterviewDashboardSummaryDto"
          },
          "outcomeSeries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyInterviewDashboardOutcomeSeriesItemDto"
            }
          },
          "weakTopics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewWeakTopicDto"
            }
          },
          "feedbackEntries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewFeedbackEntryDto"
            }
          },
          "recentInterviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewDashboardRecentItemDto"
            }
          }
        },
        "required": [
          "summary",
          "outcomeSeries",
          "weakTopics",
          "feedbackEntries",
          "recentInterviews"
        ]
      },
...
    }
  }
}